requirements.txt
# Core ML and AI libraries
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0
peft>=0.6.0
bitsandbytes>=0.41.0
datasets>=2.14.0

# vLLM for efficient inference
vllm>=0.2.0

# API and web framework
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0
openai>=1.3.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=1.0.0
pydantic>=2.5.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development tools
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# Hugging Face
huggingface-hub>=0.19.0

# Additional utilities
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# File handling
openpyxl>=3.1.0
xlsxwriter>=3.1.0 